sentinel_backend/pyproject.toml
[tool.poetry]
name = "sentinel-backend"
version = "0.1.0"
description = "Backend services for the Sentinel AI Agentic API Testing Platform"
authors = ["Cline"]

[[tool.poetry.packages]]
include = "auth_service"

[[tool.poetry.packages]]
include = "api_gateway"

[[tool.poetry.packages]]
include = "spec_service"

[[tool.poetry.packages]]
include = "orchestration_service"

[[tool.poetry.packages]]
include = "execution_service"

[[tool.poetry.packages]]
include = "data_service"

[[tool.poetry.packages]]
include = "config"

[[tool.poetry.packages]]
include = "tests"

[tool.poetry.dependencies]
python = "^3.9"

# Web Framework
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.6" # For pgvector, though asyncpg is primary
pgvector = "^0.2.0"
alembic = "^1.11.1" # For database migrations

# API Spec Parsing
prance = ">=2.0.0"
openapi-core = "^0.18.0"
pyyaml = "^6.0"

# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.2" # For testing API clients (updated for mistralai compatibility)

# Scheduling
schedule = "^1.2.0"

# LLM Providers
openai = "^1.0.0"  # Updated to latest version for better async support
anthropic = "^0.25.0"  # For Claude models
google-generativeai = "^0.5.0"  # For Gemini models
mistralai = "^0.1.0"  # For Mistral models
tiktoken = "^0.7.0"  # For OpenAI token counting
aiohttp = "^3.9.0"  # For async HTTP requests (Ollama)

# Authentication & Security
pyjwt = "^2.8.0"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"
email-validator = "^2.0.0"

# Configuration Management
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
Faker = "^22.0.0"

# Observability
structlog = "^23.1.0"
python-json-logger = "^2.0.7"
prometheus-fastapi-instrumentator = "^6.0.0"
opentelemetry-api = "^1.18.0"
opentelemetry-sdk = "^1.18.0"
opentelemetry-exporter-jaeger = "^1.18.0"
opentelemetry-instrumentation-fastapi = "^0.40b0"
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
